requirements.txt
amqp==5.0.6
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.10
importlib-metadata==3.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
numpy==1.19.5
openpyxl==3.0.7
pandas==1.1.5
prompt-toolkit==3.0.18
protobuf==3.15.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
six==1.15.0
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
xlrd==2.0.1
zipp==3.4.1
